More unit tests. Added more parameters checks.

+++ challenge/src/main/java/com/rho/services/APIInterface.java
@@ -8,6 +8,6 @@ public interface APIInterface {
     
     public JSONObject getExchangeRates(ExchangeDB db, String currency, String targets);
     public JSONObject getAllExchangeRates(ExchangeDB db, String currency);
-    public JSONObject convertCurrency(ExchangeDB db, String from, String to, int amount);
+    public JSONObject convertCurrency(ExchangeDB db, String from, String to, double amount);
     
 }
